🎨 Refine static qubit handling to require less casting

+++ mlir/lib/Conversion/QCToQCO/QCToQCO.cpp
@@ -307,8 +307,7 @@ struct ConvertQCStaticOp final : StatefulOpConversionPattern<qc::StaticOp> {
     auto qcQubit = op.getQubit();
 
     // Create new qco.static operation with the same index
-    auto qcoOp = qco::StaticOp::create(rewriter, op.getLoc(),
-                                       static_cast<int64_t>(op.getIndex()));
+    auto qcoOp = qco::StaticOp::create(rewriter, op.getLoc(), op.getIndex());
 
     // Collect QCO qubit SSA value
     auto qcoQubit = qcoOp.getQubit();
